Improve the job assets for Summer 2023 job hunt
The resume's basically good, but I wanted to fix stuff that had been
bothering me.  I added a note of my English proficiency since I'm
applying in a non-English speaking country.

The cover letter now is more oriented toward what I'm looking for from
a job rather than trying to appeal to what a company is looking for in a
candidate.  I'm doing it this way partly because it's a pain in the butt
to constantly have to rewrite parts of my cover letter for every
company, but also, it's emotionally painful to have to describe how I'm
going to be the perfect candidate for a job I frankly don't care about.

My hope is that this gambit pays off in the specificity of my desires
being appealing to those that could fulfill them.  We'll see I suppose.