Add a resume in Finnish
My Finnish homework this week was to write a CV in Finnish, so I've just
translated my normal one completely.  I also found a couple mistakes in
the English, so I've changed them as well.

I won't be publishing the Finnish version openly until it gets some
feedback, though, so no reference on the root page.