Move the call to action from the site theme to an org-mode macro
The issue with it being in the theme is that it would render below the
references section, which readers are liable to ignore.  I need it to
be after the content, but before the references, so generating the text
in org-mode is the way to go.

I wish I didn't have to use lisp code to generate it; there's already
the {{{title}}} macro, but org-mode doesn't do recursive macro
expansion, so lisp is necessary.